+++ packages/shared/lib/core/network/utils/index.ts
@@ -5,6 +5,7 @@ export * from './doesNodeHavePlugin'
 export * from './formatProtocolName'
 export * from './getDefaultClientOptions'
 export * from './getNetwork'
+export * from './getFormattedTimeStamp'
 export * from './getOfficialExplorerUrl'
 export * from './getOfficialNetwork'
 export * from './getOfficialNodes'
